Changed float32 to float16

+++ pyrat_env/wrappers.py
@@ -23,7 +23,7 @@ def flatdim2(space):
 
 def flatten2(space, x):
     if isinstance(space, Box):
-        return np.asarray(x, dtype=np.float32).flatten()
+        return np.asarray(x, dtype=np.float16).flatten()
     elif isinstance(space, Discrete):
         return [x]
     elif isinstance(space, Tuple):
@@ -43,7 +43,7 @@ class FlattenObservationMatrices(ObservationWrapper):
         super(FlattenObservationMatrices, self).__init__(env)
 
         flatdim = flatdim2(env.observation_space)
-        self.observation_space = Box(low=-float('inf'), high=float('inf'), shape=(flatdim,), dtype=np.float32)
+        self.observation_space = Box(low=-float('inf'), high=float('inf'), shape=(flatdim,), dtype=np.float16)
 
     def observation(self, observation):
         return flatten2(self.env.observation_space, observation)
@@ -54,7 +54,7 @@ class MatricizePositions(ObservationWrapper):
         super(MatricizePositions, self).__init__(env)
 
         obs_dim = (9, 21, 15)
-        self.observation_space = Box(low=-float('inf'), high=float('inf'), shape=obs_dim, dtype=np.float32)
+        self.observation_space = Box(low=-float('inf'), high=float('inf'), shape=obs_dim, dtype=np.float16)
 
     def observation(self, observation):
         ret = []
@@ -63,14 +63,14 @@ class MatricizePositions(ObservationWrapper):
         ret.append(observation['Maze_left'])
         ret.append(observation['Maze_down'])
         ret.append(observation['pieces_of_cheese'])
-        player1_score_matrix = np.full_like(observation['pieces_of_cheese'], observation['player1_score'], dtype= np.float32)
-        player2_score_matrix = np.full_like(observation['pieces_of_cheese'], observation['player2_score'], dtype= np.float32)
+        player1_score_matrix = np.full_like(observation['pieces_of_cheese'], observation['player1_score'], dtype= np.float16)
+        player2_score_matrix = np.full_like(observation['pieces_of_cheese'], observation['player2_score'], dtype= np.float16)
         ret.append(player1_score_matrix)
         ret.append(player2_score_matrix)
 
         # player location matrices
-        player1 = np.zeros(ret[0].shape, dtype= np.float32)
-        player2 = np.zeros(ret[0].shape, dtype= np.float32)
+        player1 = np.zeros(ret[0].shape, dtype= np.float16)
+        player2 = np.zeros(ret[0].shape, dtype= np.float16)
 
         player1[observation['player1_location']] = 1
         player2[observation['player2_location']] = 1
@@ -78,7 +78,7 @@ class MatricizePositions(ObservationWrapper):
         ret.append(player1)
         ret.append(player2)
 
-        return np.array(ret, dtype= np.float32)
+        return np.array(ret, dtype= np.float16)
 
 
 class FinalReward(RewardWrapper):
